requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
_sysroot_linux-64_curr_repodata_hack=3=haa98f57_10
aiohappyeyeballs=2.4.0=py311h06a4308_0
aiohttp=3.10.5=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=16.1.0=hc1eb8f0_0
attrs=24.2.0=py311h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
binutils_impl_linux-64=2.40=h5293946_0
binutils_linux-64=2.40.0=hc2dff05_1
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.4.2=py311hf4808d0_0
brotli-python=1.0.9=py311h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=py311h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=11.8=hcce14f8_3
cudatoolkit=11.8.0=h6a678d5_0
cudnn=9.3.0.75=cuda11.8
dataclasses=0.8=pyh6d0b6a4_7
datasets=3.1.0=py_0
dill=0.3.8=py311h06a4308_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=py311h5eee18b_0
fsspec=2024.6.1=py311h06a4308_0
gcc_impl_linux-64=11.2.0=h1234567_1
gcc_linux-64=11.2.0=h5c386dc_1
gflags=2.2.2=h6a678d5_1
giflib=5.2.2=h5eee18b_0
glog=0.5.0=h6a678d5_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
gxx_impl_linux-64=11.2.0=h1234567_1
gxx_linux-64=11.2.0=hc2dff05_1
huggingface_hub=0.26.2=py_0
icu=73.1=h6a678d5_0
idna=3.7=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py311h06a4308_1
joblib=1.4.2=py311h06a4308_0
jpeg=9e=h5eee18b_3
kernel-headers_linux-64=3.10.0=h57e8cba_10
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libblas=3.9.0=20_linux64_mkl
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcblas=3.9.0=20_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcurl=8.9.1=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libexpat=2.6.3=h5888daf_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-devel_linux-64=11.2.0=h1234567_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=14.2.0=h77fa898_1
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=20_linux64_mkl
libmagma=2.8.0=hfdb99dd_0
libmagma_sparse=2.8.0=h9ddd185_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.25.3=he621ea3_0
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.0=h251f7ec_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-devel_linux-64=11.2.0=h1234567_1
libstdcxx-ng=14.2.0=h4852527_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libtorch=2.4.1=cuda118_h7d43891_300
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.49.2=hb9d3cd8_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=19.1.4=h024ca30_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py311h5eee18b_0
mkl=2023.2.0=h84fe81f_50496
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.1.0=py311h5eee18b_0
multiprocess=0.70.15=py311h06a4308_0
nccl=2.23.4.1=h03a54cd_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py311h06a4308_0
numexpr=2.10.1=py311h3c60e43_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.4.0=hb9d3cd8_0
orc=2.0.1=h2d29ad5_0
packaging=24.1=py311h06a4308_0
pandas=2.2.2=py311ha02d727_0
pillow=11.0.0=py311hfdbf927_0
pip=24.2=py311h06a4308_0
progressbar2=4.5.0=pyhd8ed1ab_0
propcache=0.2.0=py311h5eee18b_0
psutil=5.9.0=py311h5eee18b_0
pyarrow=16.1.0=py311ha02d727_0
pyg=2.6.1=py311_torch_2.4.0_cu118
pyparsing=3.2.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.10=he870216_0
python-dateutil=2.9.0post0=py311h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-utils=3.9.0=pyhff2d567_1
python-xxhash=2.0.2=py311h5eee18b_1
python_abi=3.11=2_cp311
pytorch=2.4.1=cuda118_py311h1771f17_300
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py311h06a4308_0
pyyaml=6.0.2=py311h5eee18b_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2024.9.11=py311h5eee18b_0
requests=2.32.3=py311h06a4308_1
s2n=1.3.27=hdbd6064_0
safetensors=0.4.5=py311h448de89_1
scikit-learn=1.5.1=py311ha02d727_0
scipy=1.14.1=py311h08b1b3b_0
setuptools=75.1.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.7=h1b44611_2
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py311h06a4308_0
sysroot_linux-64=2.17=h57e8cba_10
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py311h92b7b1e_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.1=py311h5e3497e_1
torchaudio=2.4.1=py311_cu118
torchtriton=3.1.0=py311
torchvision=0.15.2=cuda118py311h4cc2eb7_0
tqdm=4.66.5=py311h92b7b1e_0
transformers=4.45.2=py311h06a4308_0
typing-extensions=4.11.0=py311h06a4308_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py311h06a4308_0
utf8proc=2.6.1=h5eee18b_1
wheel=0.44.0=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.0=py311h5eee18b_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.6=hc292b87_0
